# backend/src/teched/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: EdTech MVP API
  version: 1.0.0
  description: API MVP OTP-Only. Focus su velocità di sviluppo e struttura contenuti flat.
servers:
  - url: http://localhost:8080/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProgress:
      type: object
      properties:
        status: { type: string, enum: [not_started, in_progress, completed] }
        lastPosition: { type: integer, description: "Secondi nel video" }
        updatedAt: { type: string, format: date-time }

    # Dashboard
...
paths:
...
  /progress/{contentUnitId}:
    patch:
      summary: Heartbeat del video (aggiorna secondo per secondo o al completamento)
      description: Aggiorna il progresso dell'utente per un contenuto specifico. Utilizzato per tracciare la posizione nel video e lo stato di completamento.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: contentUnitId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              properties:
                status: { type: string, enum: [not_started, in_progress, completed] }
                lastPosition: { type: integer }
      responses:
        '200':
          description: Progresso aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProgress'

  # Dashboard
...