# backend/src/teched/target/generated-sources/openapi/src/main/java/com/example/teched/api/ProgressApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.teched.api;

import com.example.teched.model.ProgressContentUnitIdPatchRequest;
import java.util.UUID;
import com.example.teched.model.UserProgress;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-29T22:37:12.278941+01:00[Europe/Rome]")
@Validated
@Tag(name = "progress", description = "the progress API")
public interface ProgressApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * PATCH /progress/{contentUnitId} : Heartbeat del video (aggiorna secondo per secondo o al completamento)
     * Aggiorna il progresso dell&#39;utente per un contenuto specifico. Utilizzato per tracciare la posizione nel video e lo stato di completamento.
     *
     * @param contentUnitId  (required)
     * @param progressContentUnitIdPatchRequest  (optional)
     * @return Progresso aggiornato con successo (status code 200)
     */
    @Operation(
        operationId = "progressContentUnitIdPatch",
        summary = "Heartbeat del video (aggiorna secondo per secondo o al completamento)",
        description = "Aggiorna il progresso dell'utente per un contenuto specifico. Utilizzato per tracciare la posizione nel video e lo stato di completamento.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Progresso aggiornato con successo", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserProgress.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/progress/{contentUnitId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<UserProgress> progressContentUnitIdPatch(
        @Parameter(name = "contentUnitId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("contentUnitId") UUID contentUnitId,
        @Parameter(name = "ProgressContentUnitIdPatchRequest", description = "") @Valid @RequestBody(required = false) ProgressContentUnitIdPatchRequest progressContentUnitIdPatchRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"lastPosition\" : 6, \"status\" : \"not_started\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}